+++ mods/RecursiveIndustry/src/RecursiveIndustryIds.Power.cs
@@ -6,9 +6,6 @@ public static partial class RecursiveIndustryIds
 {
     public static partial class Power
     {
-        public static readonly StaticEntityProto.ID OrbitalPowerRelay =
-            new("RecursiveIndustry_OrbitalPowerRelay");
-
         public static readonly StaticEntityProto.ID OrbitalPowerArray =
             new("RecursiveIndustry_OrbitalPowerArray");
     }